Corrected bugs
