+++ Views/CrewTableViewCell.swift
@@ -27,6 +27,27 @@ class CrewTableViewCell: UITableViewCell {
         // Initialization code
     }
 
+    func set(crew: Crew) {
+        let timeFormat = DateFormatter()
+        timeFormat.dateFormat = "HH:mm:ss.S"
+        
+        crewName.text = crew.crewName
+        crewOarImage.image = UIImage(named: crew.picFile!)
+        crewCategory.text = crew.category
+        crewNumber.text = String(crew.crewNumber)
+        if crew.startTimeLocal != nil {
+            startTime.text = timeFormat.string(from: crew.startTimeLocal!)
+        } else {
+            startTime.text = "No time"
+        }
+        if crew.endTimeLocal != nil {
+            stopTime.text = timeFormat.string(from: crew.endTimeLocal!)
+        }else {
+            stopTime.text = "No time"
+        }
+        
+        crewTime.text = crew.elapsedTime
+    }
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
